% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{alphamis_genesum_path}
\alias{alphamis_genesum_path}
\title{cache the text file if necessary, return reference}
\usage{
alphamis_genesum_path(cache = BiocFileCache::BiocFileCache(), build = "hg38")
}
\arguments{
\item{build}{character(1) defaults to "hg38"}
}
\description{
cache the text file if necessary, return reference
}
