% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{am_gene_url}
\alias{am_gene_url}
\title{location of gene-level averages of AlphaMissense scores}
\usage{
am_gene_url(build = "hg38")
}
\arguments{
\item{build}{character(1) defaults to "hg38"}
}
\description{
location of gene-level averages of AlphaMissense scores
}
