% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{am_genemns_hg19}
\alias{am_genemns_hg19}
\title{a data.frame of gene-level mean AlphaMissense pathogenicity scores computed for hg19}
\format{
data.frame
}
\usage{
am_genemns_hg19
}
\description{
a data.frame of gene-level mean AlphaMissense pathogenicity scores computed for hg19
}
\examples{
data(am_genemns_hg19)
head(am_genemns_hg19)
}
\keyword{datasets}
