% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{am_genemns_hg38}
\alias{am_genemns_hg38}
\title{a data.frame of gene-level mean AlphaMissense pathogenicity scores computed for hg38}
\format{
data.frame
}
\usage{
am_genemns_hg38
}
\description{
a data.frame of gene-level mean AlphaMissense pathogenicity scores computed for hg38
}
\examples{
data(am_genemns_hg38)
head(am_genemns_hg38)
}
\keyword{datasets}
