% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amentgr}
\alias{amentgr}
\title{a GRanges with GWAS hits for outcomes mentioning "asthma"}
\format{
GRanges
}
\usage{
amentgr
}
\description{
a GRanges with GWAS hits for outcomes mentioning "asthma"
}
\note{
Produced on 21 Sept 2023 using `cc = gwascat::get_cached_gwascat()`, filtering
using `grep("sthma", cc$MAPPED_TRAIT)` and removing records with missing `CHR_POS`.
}
\examples{
data(amentgr)
amentgr[1:5,1:4]
}
\keyword{datasets}
