% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amnames}
\alias{amnames}
\title{a vector of column names for AlphaMissense tables}
\format{
character vector
}
\usage{
amnames
}
\description{
a vector of column names for AlphaMissense tables
}
\keyword{datasets}
