% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_alphamis_txf}
\alias{get_alphamis_txf}
\title{obtain AlphaMissense TabixFile from cache, after
retrieving the relevant resource if not present there}
\usage{
get_alphamis_txf(cache = BiocFileCache::BiocFileCache(), build = "hg38")
}
\arguments{
\item{cache}{An instance of BiocFileCache}

\item{build}{character(1) defaults to "hg38"}
}
\description{
obtain AlphaMissense TabixFile from cache, after
retrieving the relevant resource if not present there
}
\note{
First invocation on a given R deployment will
a) request that a cache be built if BiocFileCache has
never been used, b) populate the cache with ~600MB
of bgzipped tabix-indexed variant-level data.
}
\examples{
ca = BiocFileCache::BiocFileCache()
txf = get_alphamis_txf(ca) # will download if data are not already cached
Rsamtools::yieldSize(txf) = 10L
read.delim(text=Rsamtools::scanTabix(txf)[[1]], h=FALSE) 
}
