% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{osn_alphamis_url}
\alias{osn_alphamis_url}
\title{produce URLs for AlphaMissense SNV results lodged in NSF Open Storage Network}
\usage{
osn_alphamis_url(build = "hg38")
}
\arguments{
\item{build}{character(1) defaults to "hg38"}
}
\description{
produce URLs for AlphaMissense SNV results lodged in NSF Open Storage Network
}
\examples{
osn_alphamis_url()
}
